[#40] fix: 오류해결
