+++ src/components/mypage/MyPageLayout.tsx
@@ -1,10 +1,10 @@
 'use client';
 import Link from 'next/link';
-import useProfile from './ProfileManager';
 import StampNum from './StampNum';
+import useUserNickname from '@/hooks/useUserNickname';
 
 const MyPageLayout = () => {
-  const { nickname } = useProfile(); // 프로필 정보와 로그아웃 기능 가져오기
+  const { nickname, error } = useUserNickname();
 
   return (
     <div className="flex flex-col items-start justify-center p-6">
